server/pyproject.toml
[project]
name = "ace-server"
version = "0.1.0"
description = "ACE coordination server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
    "ollama>=0.5.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
